Add the ability to customize indices to render to the screen
This is the original idea i have when creating this new Plane class. I
was thinking that instead of rendering each cell as its own square, I
can modify the indices, so the memory overhead and (probably) speed will
be smaller.

I don't know why on previous commit I use the texture update technique.
It defeats the very purpose of the creation of this class itself.