Fix and move threading, also fix orientation inconsistency
Orientation on the program previously was flipped on y-axis. This makes
operation confusing and many flipping done here and there. This leads to
consistency. On this one, I fix y-axis orientation so that the positive y-axis
points to the up direction.

On the threading topic, on previous commit, I forgot to remove a piece
of code that corresponds to a function that updates the indices (see
`RenderEngine::drawGrid`, there `setState` was called).Updating indices
should be in a separate thread. This issue now resolved.

The threading capability is moved from GridTile to Plane, this makes it
easier preventing data race. Data race mainly avoided using mutex and
two indices data: one used for drawing (`Plane::m_indices_custom_render`), and
one used to be modified (`Plane::m_indices_custom_edit`). These indices
data then swapped once in a while.